application/physical/python/requirements.txt
google-cloud-storage==1.26.0
Flask==1.1.1
requests==2.23.0
urllib3==1.25.8
gunicorn==19.3.0